Check the enabled status for returned hash

+++ safelist.py
@@ -46,22 +46,23 @@ class Safelist(ServiceBase):
             resp = self.session.get(f"{self.service_api_host}/api/v1/safelist/{qhash}/")
             if resp.ok:
                 data = resp.json()['api_response']
-                # Create a section per source
-                for source in data['sources']:
-                    if source['type'] == 'user':
-                        msg = f"User {source['name']} deemed this file as safe for the following reason(s):"
-                        heur_id = 2
-                    else:
-                        msg = f"External safelist source {source['name']} deems this file as safe " \
-                            "for the following reason(s):"
-                        heur_id = 1
+                if data['enabled'] and data['type'] == "file":
+                    # Create a section per source
+                    for source in data['sources']:
+                        if source['type'] == 'user':
+                            msg = f"User {source['name']} deemed this file as safe for the following reason(s):"
+                            heur_id = 2
+                        else:
+                            msg = f"External safelist source {source['name']} deems this file as safe " \
+                                "for the following reason(s):"
+                            heur_id = 1
 
-                    result.add_section(
-                        ResultSection(
-                            msg, heuristic=Heuristic(heur_id, signature=f"SAFELIST_{qhash}"),
-                            body="\n".join(source['reason'])))
+                        result.add_section(
+                            ResultSection(
+                                msg, heuristic=Heuristic(heur_id, signature=f"SAFELIST_{qhash}"),
+                                body="\n".join(source['reason'])))
 
-                # Stop processing, the file is safe
-                request.drop()
+                    # Stop processing, the file is safe
+                    request.drop()
 
         request.result = result
